backend/requirements.txt
APScheduler==3.11.1
asgiref==3.10.0
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
cryptography==46.0.3
Django==5.2.8
django-allauth==65.13.0
django-cors-headers==4.9.0
django-restframework==0.0.1
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
idna==3.11
logger==1.4
pillow==12.0.0
pycparser==2.23
PyJWT==2.10.1
requests==2.32.5
shortuuid==1.0.13
sqlparse==0.5.3
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.5.0
